Shooting added.
Creating a new Texture anywhere but in the GLSurfaceView threw an Exception.. Why?
Because OpenGL calls can only be made on UI thread, and to do that from now on i shall use "glSurfaceView?.queueEvent({ /*Code comes here*/ })"
It took quite a lot of time to come to this conclusion. More than what i feel i should've spent on it.